requirements.txt
selenium==4.15.0
pytest==7.4.0
allure-pytest==2.13.0
requests==2.31.0